Chore: Update to lit 3.x.x
Includes switching to level 3  decorators in stead of  experimental TypeScript decorators as babel does not support experimental TypeScript decorators.

To ensure that the new style decorators are also recognized by the esbuild used by the web-dev, I've overridden the esbuild version that is required by @web/dev-server-esbuild to be at least 0.24, such that decorators level 3 are supported.

For level 3 decorators, decorated fields need to be accessor field. Added accessor to all such fields.

Updated babel configuration to alles transform class static blocks, which is needed for the decorators.